+++ HISTORY.md
@@ -27,6 +27,8 @@ Our backwards-compatibility policy can be found [here](https://github.com/python
 - Many preconf converters (_bson_, stdlib JSON, _cbor2_, _msgpack_, _msgspec_, _orjson_, _ujson_) skip unstructuring `int` and `str` enums,
   leaving them to the underlying libraries to handle with greater efficiency.
   ([#598](https://github.com/python-attrs/cattrs/pull/598))
+- The {class}`msgspec JSON preconf converter <cattrs.preconf.msgspec.MsgspecJsonConverter>` now handles dataclasses with private attributes more efficiently.
+  ([#624](https://github.com/python-attrs/cattrs/pull/624))
 - Literals containing enums are now unstructured properly, and their unstructuring is greatly optimized in the _bson_, stdlib JSON, _cbor2_, _msgpack_, _msgspec_, _orjson_ and _ujson_ preconf converters.
   ([#598](https://github.com/python-attrs/cattrs/pull/598))
 - Preconf converters now handle dictionaries with literal keys properly.